The Si4707 defines this as uint32_t so use the same here. This may go away as the library defines a bunch of globals we can use for our EEPROM stuff.

+++ WeatherRadio.ino
@@ -33,7 +33,7 @@ int configAddress = 0;
 struct StoreStruct
 {
   char version[4];
-  int frequency;
+  unsigned long frequency;
 } storage;
